Finishing Up the required features

+++ src/Kawadar.Application/Features/Badges/Commands/DeleteBadge/DeleteBadgeCommandHandler.cs
@@ -2,13 +2,15 @@
 using Kawadar.Application.Common.Interfaces.Auth;
 using Kawadar.Application.Common.Interfaces.Repositories;
 using Kawadar.Domain.Badges;
+using Kawadar.Domain.Common.Constants;
 using Kawadar.Domain.Common.Results;
+using Kawadar.Domain.StorageRepository;
 using MediatR;
 
 
 namespace Kawadar.Application.Features.Badges.Commands.DeleteBadge
 {
-    public class DeleteBadgeCommandHandler(IUnitOfWork unitOfWork,IUser user, IBadgeRepository badgeRepository) : IRequestHandler<DeleteBadgeCommand, Result<Deleted>>
+    public class DeleteBadgeCommandHandler(IUnitOfWork unitOfWork,IUser user, IBadgeRepository badgeRepository, IStorageClient storageClient) : IRequestHandler<DeleteBadgeCommand, Result<Deleted>>
     {
         public async Task<Result<Deleted>> Handle(DeleteBadgeCommand request, CancellationToken cancellationToken)
         {
@@ -18,6 +20,10 @@ namespace Kawadar.Application.Features.Badges.Commands.DeleteBadge
             var result = await badgeRepository.GetById(request.badgeId);
             if (result.IsError) return result.Errors;
             var badge = result.Value;
+            var storageDeleteResult = await storageClient.DeleteFileAsync(badge.IconUrl, Containers.Badges);
+
+            if (storageDeleteResult.IsError) return storageDeleteResult.Errors;
+
             var deleteResult = badgeRepository.Delete(badge);
             if (deleteResult.IsError) return deleteResult.Errors;
             await unitOfWork.SaveChangesAsync(cancellationToken);
